Fix so that C++ compiler don't break on our header. Include extern "C" in most place, fix some non conforming return type, rename "new" / "operator" variable.
git-svn-id: file:///home/yoann/SVN/trunk/libprelude@5704 09c5ec92-17d4-0310-903a-819935f44dba

+++ src/include/prelude-extract.h
@@ -80,7 +80,7 @@ static inline uint64_t prelude_align_uint64(const void *buf)
 
 static inline float prelude_align_float(const void *buf)
 {
-        return (float) prelude_align_uint32(buf);
+        return prelude_align_uint32(buf);
 }
